pyproject.toml
[tool.black]
line-length = 80
target-version = ['py312']
include = '\.pyi?$'
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | foo.py           # also separately exclude a file named foo.py in
                     # the root of the project
)
'''

[tool.interrogate]
ignore-init-method = true
ignore-init-module = false
ignore-magic = false
ignore-semiprivate = false
ignore-private = false
ignore-property-decorators = false
ignore-module = true
fail-under = 100
exclude = [
  "setup.py",
  "docs",
  "build",
  "data",
  "bin",
  "dandelion/external",
  "tests",
]
ignore-regex = ["^get$", "^mock_.*", ".*BaseClass.*"]
verbose = 2
quiet = false
whitelist-regex = []
color = true

[tool.pytest.ini_options]
filterwarnings = [
  "ignore::DeprecationWarning",
  "ignore::PendingDeprecationWarning",
  "ignore::FutureWarning",
]
addopts = "--cov src/dandelion --cov-report=xml -rP"
testpaths = ["tests"]

[tool.coverage.run]
omit = ["containter/*", "src/dandelion/external/*", "src/dandelion/tutorial/*"]

[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.0"]
build-backend = "setuptools.build_meta"

[project]
name = "sc-dandelion"
dynamic = ["version"]
authors = [{ name = "zktuong", email = "z.tuong@uq.edu.au" }]
description = "sc-TCR/BCR-seq analysis tool"
readme = { file = "README.md", content-type = "text/markdown" }
license = { text = "GNU Affero General Public License v3 or later (AGPLv3+)" }
classifiers = [
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
]
keywords = ["dandelion", "single-cell", "immune", "TCR", "BCR"]
dependencies = [
  "airr>=1.4",
  "setuptools_scm[toml]>=6.0",
  "numpy>=1.23",
  "pandas>=1.4",
  "changeo>=1.1",
  "anndata>=0.9",
  "scanpy>=1.9",
  "scikit-learn>=1.0",
  "scipy>=1.8",
  "numba>=0.56",
  "seaborn>=0.12",
  "networkx>=2.8",
  "leidenalg>=0.9",
  "polyleven>=0.4",
  "h5py>=3.6",
  "adjustText>=0.7",
  "distance>=0.1.3",
  "plotnine>=0.10",
  "palettable>=3.3",
  "mizani>=0.8",
  "nxviz>=0.7",
]

[project.urls]
homepage = "https://github.com/zktuong/dandelion/"
documentation = "https://sc-dandelion.readthedocs.io/"

[project.optional-dependencies]
docs = [
  "airr>=1.4",
  "ipykernel>=6.0",
  "biopython>=1.79",
  "changeo>=1.1",
  "presto>=0.7",
  "pyyaml>=5.4",
  "yamlordereddictloader>=0.4",
  "sphinx>=4.5",
  "nbsphinx>=0.8",
  "sphinx-autodoc-typehints>=1.18",
  "sphinx_rtd_theme>=1.0",
  "readthedocs-sphinx-ext>=2.1",
  "recommonmark>=0.7",
]
scirpy = ["scirpy>=0.21", "awkward>=2.1", "mudata>=0.2"]
scrublet = ["scrublet>=0.2", "annoy>=1.16"]
palantir = ["palantir>=1.0", "pertpy>=1.0", "jax>=0.3"]

[tool.setuptools]
include-package-data = true
package-data = { "dandelion" = [
  "bin/tigger-genotype.R",
  "bin/MakeDb_gentle.py",
] }
# Use data-files for extra files, if required
data-files = { bin = ["bin/tigger-genotype.R", "bin/MakeDb_gentle.py"] }

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools_scm]
version_scheme = "guess-next-dev"
local_scheme = "dirty-tag"
